tidied up the components by moving the handle functions into the respected components
