+++ data/db.json
@@ -23,12 +23,6 @@
       "title": "My Fourth Blog",
       "datetime": "07 August 2024 1:13:40 PM",
       "body": "HI!"
-    },
-    {
-      "id": "41",
-      "title": "fviuvhui",
-      "datetime": "14 August 2024 3:37:31 PM",
-      "body": "duvhduvheuddvnfjvn"
     }
   ]
 }
