+++ src/EditPost.js
@@ -1,13 +1,19 @@
-import { useEffect, useContext } from 'react'
-import { useParams, Link } from 'react-router-dom'
+import { useState, useEffect, useContext } from 'react'
+import { useParams, Link, useNavigate } from 'react-router-dom'
 import DataContext from './context/DataContext';
+import api from './api/posts';
+import { format } from 'date-fns';
 
 
 
 const EditPost = () => {
-    const { posts, handleEdit, editBody, setEditBody, editTitle, setEditTitle } = useContext(DataContext)
+    const [editTitle, setEditTitle] = useState('')
+    const [editBody, setEditBody] = useState('')
+    const { posts, setPosts } = useContext(DataContext)
     const { id } = useParams()
     const post = posts.find(post => (post.id).toString() === id) 
+    const navigate = useNavigate()
+    
 
     useEffect(() => {
         if (post) {
@@ -15,6 +21,21 @@ const EditPost = () => {
             setEditBody(post.body)
         }
     }, [post, setEditTitle, setEditBody])
+
+    const handleEdit = async (id) => {
+        const datetime = format(new Date(), 'dd MMMM yyyy pp');
+        const updatePost = { id, title: editTitle, datetime, body: editBody }
+        try {
+          const response = await api.put(`/posts/${id}`, updatePost)
+          setPosts(posts.map(post => post.id === id ? { ...response.data } : post))
+          setEditTitle('')
+          setEditBody('')
+          navigate('/')
+        } catch(err) {
+          console.log(`Error: ${err.message}`)
+        }
+      }
+    
 
     return (
        <main className="New Post">
