+++ src/Header.js
@@ -1,9 +1,8 @@
 import { FaLaptop, FaTabletAlt, FaMobileAlt } from 'react-icons/fa'
-import { useContext } from 'react';
-import DataContext from './context/DataContext';
+import useWindowSize from './Hooks/useWindowSize'
 
 const Header = ({ title }) => {
-    const { width } = useContext(DataContext)
+    const { width } = useWindowSize()
     return (
         <header className="Header">
             <h1>{title}</h1>
