+++ src/NewPost.js
@@ -1,8 +1,34 @@
-import { useContext } from 'react';
+import { useState, useContext } from 'react';
 import DataContext from './context/DataContext';
+import { useNavigate } from 'react-router-dom'
+import api from './api/posts'
+import { format } from 'date-fns';
 
 const NewPost = () => {
-    const { handleSubmit, postTitle, setPostTitle, postBody, setPostBody } = useContext(DataContext)
+    const [postTitle, setPostTitle] = useState('')
+    const [postBody, setPostBody] = useState('')
+    const { posts, setPosts } = useContext(DataContext)
+    const navigate = useNavigate()
+
+    const handleSubmit = async (e) => {
+        e.preventDefault()
+        const id = posts.length ? posts[posts.length - 1].id + 1 : 1;
+        const datetime = format(new Date(), 'dd MMMM yyyy pp');
+        const newPost = { id, title: postTitle, datetime, body: postBody }
+        
+
+        try {
+          const response = await api.post('/posts', newPost)
+          const allPosts = [ ...posts, response.data ]
+          setPosts(allPosts);
+          setPostTitle('');
+          setPostBody('');
+          navigate('/');
+        } catch(err) {
+          console.log(`Error: ${err.message}`)
+        }
+      }
+
     return (
         <main className="New Post">
             <h2>New Post</h2>
