+++ src/PostPage.js
@@ -1,8 +1,25 @@
-import { useParams, Link } from 'react-router-dom'
+import { useParams, Link, useNavigate } from 'react-router-dom'
+import { useContext } from 'react';
+import api from './api/posts';
+import DataContext from './context/DataContext';
 
-const PostPage = ({ posts, handleDelete }) => {
+const PostPage = () => {
+    const { posts, setPosts } = useContext(DataContext)
     const { id } = useParams();
     const post = posts.find(post => (post.id).toString() === id)
+    const navigate = useNavigate()
+
+    const handleDelete = async (id) => {
+        try {
+          await api.delete(`/posts/${id}`)
+          const postsList = posts.filter(post => post.id !== id)
+          setPosts(postsList)
+          navigate('/')
+        } catch(err) {
+          console.log(`Error: ${err.message}`)
+        }
+      }
+
     return (
         <main className="PostPage">
             <article className="post">
