+++ src/context/DataContext.js
@@ -1,8 +1,4 @@
 import { createContext, useEffect, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { format } from 'date-fns';
-import api from '../api/posts'
-import useWindowSize from '../Hooks/useWindowSize';
 import useAxiosFetch from '../Hooks/useAxiosFetch';
 
 const DataContext = createContext({})
@@ -12,13 +8,6 @@ export const DataProvider = ({children}) => {
     const [posts, setPosts] = useState([])
     const [search, setSearch] = useState("");
     const [searchResults, setSearchResults] = useState([]);
-    const [postTitle, setPostTitle] = useState('')
-    const [postBody, setPostBody] = useState('')
-    const [editTitle, setEditTitle] = useState('')
-    const [editBody, setEditBody] = useState('')
-    const navigate = useNavigate();
-    const { width } = useWindowSize()
-
     const { data, fetchError, isLoading } = useAxiosFetch('http://localhost:3500/posts')
 
   useEffect(() => {
@@ -33,54 +22,11 @@ export const DataProvider = ({children}) => {
       setSearchResults(filteredResults.reverse())
   }, [posts, search])
 
-  const handleSubmit = async (e) => {
-    e.preventDefault()
-    const id = posts.length ? posts[posts.length - 1].id + 1 : 1;
-    const datetime = format(new Date(), 'dd MMMM yyyy pp');
-    const newPost = { id, title: postTitle, datetime, body: postBody }
-    try {
-      const response = await api.post('/posts', newPost)
-      const allPosts = [ ...posts, response.data ]
-      setPosts(allPosts);
-      setPostTitle('');
-      setPostBody('');
-      navigate('/');
-    } catch(err) {
-      console.log(`Error: ${err.message}`)
-    }
-  }
-
-  const handleDelete = async (id) => {
-    try {
-      await api.delete(`/posts/${id}`)
-      const postsList = posts.filter(post => post.id !== id)
-      setPosts(postsList)
-      navigate('/')
-    } catch(err) {
-      console.log(`Error: ${err.message}`)
-    }
-  }
-
-  const handleEdit = async (id) => {
-    const datetime = format(new Date(), 'dd MMMM yyyy pp');
-    const updatePost = { id, title: editTitle, datetime, body: editBody }
-    try {
-      const response = await api.put(`/posts/${id}`, updatePost)
-      setPosts(posts.map(post => post.id === id ? { ...response.data } : post))
-      setEditTitle('')
-      setEditBody('')
-      navigate('/')
-    } catch(err) {
-      console.log(`Error: ${err.message}`)
-    }
-  }
-
     return (
         <DataContext.Provider value={{
-            width, search, setSearch, 
+            search, setSearch, 
             searchResults, fetchError, isLoading, 
-            handleSubmit, postTitle, setPostTitle, postBody, setPostBody,
-            posts, handleEdit, editBody, setEditBody, editTitle, setEditTitle
+            posts, setPosts
         }}>
             {children}
         </DataContext.Provider>
